requirements.txt
torch==2.0.0
torchvision==0.15.0
numpy==1.23.5
matplotlib==3.6.2
Pillow==9.4.0
opencv-python==4.7.0.72
